refactor worker
